+++ MagnetarCraft/include/starcraft.h
@@ -37,6 +37,7 @@ DEFINE_ENUM_FLAG_OPERATORS(SaiAccessabilityFlags);
 DEFINE_ENUM_FLAG_OPERATORS(PrintFlags);
 DEFINE_ENUM_FLAG_OPERATORS(StatusFlags);
 
+extern bool end_mission_prompt;
 extern bool keep_app_active_in_background;
 
 void localDll_Init_(HINSTANCE a1);