+++ CHANGELOG.md
@@ -4,7 +4,9 @@
   - Add a new trigger action: "Execute Lua"
   - Fix an issue in melee unit initialization (introduced in 0.2.0)
   - Fix an issue in sprite initialization (introduced in 0.2.0)
-  - Add a command line argument `-keep-app-active` that keeps the app active in the background
+  - Add command line arguments:
+    * `-keep-app-active` - keeps the app active in the background
+    * `-skip-end-mission-prompt` - remove the win/lose dialog upon mission completion
 
 ## 0.2.0 - Busy Barracks
   - Reimplement ~10% of the functions in StarCraft 1.16.1